still not working....

+++ gudpy/core/modulation_excitation.py
@@ -10,7 +10,7 @@ from core import config
 from copy import deepcopy
 import tempfile
 import shutil
-
+from watchpoints import watch
 
 SUFFIX = ".exe" if os.name == "nt" else ""
 
@@ -73,6 +73,7 @@ class ModulationExcitation():
         self.startPulse = None
         self.auxDir = None
         self.outputDir = None
+        watch(self.outputDir)
         self.sample = None
         self.useDefinedPulses = True
         self.tmp = tempfile.TemporaryDirectory()
@@ -93,6 +94,10 @@ class ModulationExcitation():
                 ), f"modulation_excitation{SUFFIX}"
             )
         if useTempDir:
+            spec_bad = os.path.join(
+                self.ref.instrument.GudrunInputFileDir,
+                "spec.bad"
+            )
             if headless:
                 for dataFile in self.ref.sampleBackgrounds[0].samples[0].dataFiles.dataFiles:
                     shutil.copyfile(
@@ -105,6 +110,13 @@ class ModulationExcitation():
                             dataFile
                         )
                     )
+                    shutil.copyfile(
+                        spec_bad,
+                        os.path.join(
+                            self.tmp.name,
+                            "spec.bad"
+                        )
+                    )
             else:
                 for dataFile in self.ref.normalisation.dataFiles.dataFiles:
                     tasks.append((shutil.copyfile, [os.path.join(self.ref.instrument.dataFileDir, dataFile), os.path.join(self.tmp.name, dataFile)]))
@@ -113,6 +125,7 @@ class ModulationExcitation():
                 for container in self.ref.sampleBackgrounds[0].samples[0].containers:
                     for dataFile in container.dataFiles.dataFiles:
                         tasks.append((shutil.copyfile, [os.path.join(self.ref.instrument.dataFileDir, dataFile), os.path.join(self.tmp.name, dataFile)]))
+                tasks.append((shutil.copyfile, [spec_bad, os.path.join(self.tmp.name, "spec.bad")]))
         if headless:
             self.write_out()
             result = subprocess.run(
@@ -129,7 +142,10 @@ class ModulationExcitation():
 
     def copyfile(self, src, dest):
         print(f"Copying {src} to {dest}")
-        shutil.copyfile(src, dest)
+        if os.path.exists(src):
+            shutil.copyfile(src, dest)
+        else:
+            print(f"{src} doesn't exist.")
 
     def process(self, files, headless=True):
         tasks = []
@@ -139,11 +155,11 @@ class ModulationExcitation():
             gf.instrument.dataFileDir = self.tmp.name + "/"
             gf.sampleBackgrounds[0].samples[0].dataFiles.dataFiles = [base]
             gf.instrument.GudrunInputFileDir = self.tmp.name
-            base = os.path.split(f)[0]
+            base = os.path.splitext(f)[0]
             print("BASE: " + base)
             if headless:
                 gf.process()
-                shutil.copyfile(os.path.join(self.tmp.name, base+".mint01"), os.path.join(self.outputDir, base+".mint01"))
+                self.copyfile(os.path.join(self.tmp.name, base+".mint01"), os.path.join(self.outputDir, base+".mint01"))
             else:
                 dcs, func, args = gf.dcs(
                     path = os.path.join(
@@ -157,7 +173,7 @@ class ModulationExcitation():
                 # tasks.append(gf.dcs(path=os.path.join(self.tmp.name, "gudpy.txt"), headless=False),  + (self.tmp.name))
                 src = os.path.join(self.tmp.name, base+".mint01")
                 dest = os.path.join(self.outputDir, base+".mint01")
-                tasks.append((self.copyfile, [src, dest]))
+                tasks.append((self.copyfile, [src, dest,]))
         if not headless:
             return tasks
         # for f in files:
@@ -192,8 +208,8 @@ class ModulationExcitation():
 
         return (
             f"{self.tmp.name}\n"
-            f"{os.path.join(self.gudrunFile.instrument.GudrunStartFolder, self.gudrunFile.instrument.nxsDefinitionFile)}\n"
-            f"{len(self.gudrunFile.sampleBackgrounds[0].samples[0].dataFiles.dataFiles)}\n"
+            f"{os.path.join(self.ref.instrument.GudrunStartFolder, self.ref.instrument.nxsDefinitionFile)}\n"
+            f"{len(self.ref.sampleBackgrounds[0].samples[0].dataFiles.dataFiles)}\n"
             f"{dataFilesLines}\n"
             f"{ExtrapolationModes(self.extrapolationMode.value).name}\n"
             f"{str(self.period)}"
